faet : update component 분리
