+++ 신동현/src/pages/index.tsx
@@ -6,8 +6,9 @@ import KakaoLoginLoading from "./KakaoLoginLoading";
 import Detail from "./Detail/Detail";
 import Write from "./Write/Write";
 import WriteForm from "./Write/WriteForm";
-import Update from "./Update";
+import Update from "./Update/Update";
+import UpdateForm from "./Update/UpdateForm";
 import Root from "./Root";
 import NotFound from "./Not Found";
 
-export { Home, SignUp, SignUpDetail, Mypage, KakaoLoginLoading, Detail, Write, WriteForm, Update, Root, NotFound };
+export { Home, SignUp, SignUpDetail, Mypage, KakaoLoginLoading, Detail, Write, WriteForm, Update, UpdateForm, Root, NotFound };
